requirements.txt
django
djangorestframework
django-cors-headers
django-environ
whitenoise
